RasaChatbot/requirements.txt
pymupdf
beautifulsoup4
pinecone
sentence-transformers
rasa[full]
python-dotenv